requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.4
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
environs==9.5.0
inflection==0.5.1
marshmallow==3.20.1
packaging==23.1
phonenumbers==8.13.18
psycopg==3.1.10
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
